ピアノロール上に音符を描く処理を追加

+++ MidiInput/PianorollItem.h
@@ -9,7 +9,7 @@ public:
     /**
      * tick 単位の時刻
      */
-    int tick;
+    long int tick;
 
     /**
      * ノート番号
